feat: add a basic structure
