+++ lib/main.dart
@@ -1,5 +1,18 @@
 import 'package:bookido/app.dart';
-import 'package:bookido/app_state_container.dart';
 import 'package:flutter/material.dart';
+import 'package:bookido/models/app_state.dart';
+import 'package:scoped_model/scoped_model.dart';
 
-void main() => runApp(new AppStateContainer(child: AppRootWidget()));
+void main() {
+  // initialize the model
+  final state = new AppModel();
+
+  // run flutter UI
+  runApp(
+    ScopedModel<AppModel>(
+      model: state,
+      child: AppRoot()
+    ),
+  );
+  
+}
