+++ pubspec.yaml
@@ -27,6 +27,8 @@ dependencies:
   image_picker: ^0.4.12+1
   firebase_ml_vision: ^0.6.0+2
 
+  scoped_model: ^1.0.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
